% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfiles.r
\name{A0_First_Steps_in_R}
\alias{A0_First_Steps_in_R}
\title{A0 First Steps in R}
\description{
An introduction to using R, with information on how to use scripts, the console,
and RStudio.
}
\examples{
\dontrun{
learnr::run_tutorial("A0_First_Steps_in_R", package = "keeleR")
}
}
\keyword{Introduction}
