% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfiles.r
\name{A1_Comparing_between_groups}
\alias{A1_Comparing_between_groups}
\title{A1 Comparing between groups}
\description{
An introduction to using R for t-tests and ANOVA, Wilcoxon and Kruskal-Wallis tests.
}
\examples{
\dontrun{
learnr::run_tutorial("A1_Comparing_between_groups", package = "keeleR")
}
}
\keyword{R}
