% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfiles.r
\name{A2_Relationships_between_variables}
\alias{A2_Relationships_between_variables}
\title{A2 Relationships and associations between variables}
\description{
An introduction to using R for correlations (Pearson, Spearman) and tests of association (Chi-square).
}
\examples{
#' \dontrun{
learnr::run_tutorial("A2_Relationships_between_variables", package = "keeleR")
}
}
\keyword{R}
