% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfiles.r
\name{A3_Introduction_to_regression}
\alias{A3_Introduction_to_regression}
\title{A3 Introduction to regression}
\description{
An introduction to using R for simple linear regression.
}
\examples{
\dontrun{
learnr::run_tutorial("A3_Introduction_to_regression", package = "keeleR")
}
}
\keyword{R}
