% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfiles.r
\name{A4_Extending_regression}
\alias{A4_Extending_regression}
\title{A4 Extending regression}
\description{
An introduction to using R for multiple linear regression and other families
}
\examples{
\dontrun{
learnr::run_tutorial("A4_Extending_regression", package = "keeleR")
}
}
\keyword{R}
