% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contents.r
\name{contents}
\alias{contents}
\title{Tutorial contents}
\usage{
contents()
}
\description{
This package is primarily designed to house R tutorials using the \code{learnR} package
This function will print a list of all the tutorials currently included in the package.
}
\details{
There are no arguments.
}
