% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfiles.r
\name{GIS1_Intro_to_GIS_in_R}
\alias{GIS1_Intro_to_GIS_in_R}
\title{Introduction to GIS in R}
\description{
A brief intro to using R as a GIS
}
\examples{
\dontrun{
learnr::run_tutorial("GIS1_Intro_to_GIS_in_R", package = "keeleR")
}
}
\keyword{GIS,}
\keyword{Geographic}
\keyword{Information}
\keyword{Systems,}
\keyword{analysis}
\keyword{data}
\keyword{spatial}
