% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfiles.r
\name{GIS2_Using_rasters_in_R}
\alias{GIS2_Using_rasters_in_R}
\title{Using rasters in R}
\description{
A brief intro to using R for raster analysis
}
\examples{
\dontrun{
learnr::run_tutorial("GIS2_Using_rasters_in_R", package = "keeleR")
}
}
\keyword{GIS,}
\keyword{Geographic}
\keyword{Information}
\keyword{Systems,}
\keyword{analysis,}
\keyword{data}
\keyword{raster}
\keyword{spatial}
