% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfiles.r
\name{GIS3_Making_maps_in_R}
\alias{GIS3_Making_maps_in_R}
\title{Making maps in R}
\description{
A brief intro to making maps in R
}
\examples{
\dontrun{
learnr::run_tutorial("GIS3_Making_maps_in_R", package = "keeleR")
}
}
\keyword{GIS,}
\keyword{Geographic}
\keyword{Information}
\keyword{Systems,}
\keyword{cartography}
