% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfiles.r
\name{M1_Intro_to_multivariate}
\alias{M1_Intro_to_multivariate}
\title{Intro to multivariate data in R}
\description{
An introduction to multivariate analysis in R and RStudio.
}
\examples{
\dontrun{
learnr::run_tutorial("M1_Intro_to_multivariate", package = "keeleR")
}
}
\keyword{analysis}
\keyword{multivariate}
