% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openScript.R
\name{openScript}
\alias{openScript}
\title{Open a script in your RStudio session}
\usage{
openScript(script)
}
\arguments{
\item{script}{Character vector with name of R script in \code{extdata} folder in
package or full path to existing r script.}
}
\value{
No value - side effect of opening a script in default application.
}
\description{
openScript - utility to open practice scripts in user session. Only used in tutorials.
}
